반복설정

+++ lib/todo.dart
@@ -10,6 +10,8 @@ class Todo {
   TimeOfDay? alarmTime;
   String? repeat;
   bool isDone;
+  List<DateTime>? futureDates;
+  List<String> repeatDays;
 
   Todo({
     this.id = '',
@@ -19,6 +21,8 @@ class Todo {
     this.alarmTime,
     this.repeat,
     this.isDone = false,
+    this.futureDates,
+    this.repeatDays = const [],
   });
 
   // Firestore에서 가져온 데이터를 Map으로 변환
@@ -34,7 +38,6 @@ class Todo {
       'isDone': isDone,
     };
   }
-
 
   // Firestore의 데이터를 Todo 객체로 변환
   static Todo fromFirestore(DocumentSnapshot doc) {
@@ -95,6 +98,7 @@ class _TodoPageState extends State<TodoPage> {
       print("Error updating todo: $e");
     }
   }
+
   // Firestore에서 할 일 목록 읽기
   Future<void> _loadTodos() async {
     final todoCollection = FirebaseFirestore.instance.collection('todos');
@@ -109,13 +113,14 @@ class _TodoPageState extends State<TodoPage> {
     widget.onTodoListChanged(_todoList); // 상위 위젯에 변경 알림
   }
 
+
   // Firestore에 할 일 추가
   Future<void> _addTodoToFirestore(Todo todo) async {
     final todoCollection = FirebaseFirestore.instance.collection('todos');
     try {
       DocumentReference docRef = await todoCollection.add(todo.toMap());
       setState(() {
-        todo.id = docRef.id;  // Firestore에서 생성된 ID를 Todo에 저장
+        todo.id = docRef.id; // Firestore에서 생성된 ID를 Todo에 저장
       });
     } catch (e) {
       print("Error adding todo: $e");
@@ -132,9 +137,11 @@ class _TodoPageState extends State<TodoPage> {
     }
   }
 
+
   @override
   Widget build(BuildContext context) {
-    final filteredTodoList = _todoList.where((todo) => isSameDay(todo.date, widget.selectedDate)).toList();
+    final filteredTodoList = _todoList.where((todo) =>
+        isSameDay(todo.date, widget.selectedDate)).toList();
 
     return Scaffold(
       appBar: AppBar(
@@ -144,12 +151,15 @@ class _TodoPageState extends State<TodoPage> {
         itemCount: filteredTodoList.length,
         itemBuilder: (context, index) {
           final todo = filteredTodoList[index];
+          // 각 Todo 항목을 Dismissible 위젯으로 감싸 스와이프로 삭제 가능하게 함
           return Dismissible(
-            key: Key(todo.id),
+            key: Key(todo.id),//key: UniqueKey(),
             direction: DismissDirection.endToStart,
             onDismissed: (direction) {
               setState(() {
                 _todoList.remove(todo);
+                _generateFutureDates(todo);
+                _addFutureTodos(todo);
               });
               widget.onTodoRemoved(todo); // 할 일 삭제 콜백 호출
               _deleteTodoFromFirestore(todo.id); // Firestore에서 할 일 삭제
@@ -198,11 +208,10 @@ class _TodoPageState extends State<TodoPage> {
         ),
         Checkbox(
           value: todo.isDone,
-          onChanged: (bool? value) async {
+          onChanged: (bool? value) {
             setState(() {
               todo.isDone = value ?? false;
             });
-            await _updateTodoInFirestore(todo);
             _updateTodoList();
           },
         ),
@@ -308,8 +317,8 @@ class _TodoPageState extends State<TodoPage> {
         onTap: () => _selectAlarmTime(context, todo),
         child: _buildOptionBox(
           icon: Icons.alarm,
-          text: todo.alarmTime != null ? '${todo.alarmTime!.hour}:${todo.alarmTime!.minute}' : '알림 시간',
-
+          text: todo.alarmTime != null ? '${todo.alarmTime!.hour}:${todo
+              .alarmTime!.minute}' : '알림 시간',
         ),
       ),
     );
@@ -324,7 +333,6 @@ class _TodoPageState extends State<TodoPage> {
           icon: Icons.repeat,
           text: todo.repeat ?? '반복 설정',
         ),
-
       ),
     );
   }
@@ -338,47 +346,111 @@ class _TodoPageState extends State<TodoPage> {
     if (newTime != null) {
       setState(() {
         todo.alarmTime = newTime;
-        _updateTodoList();
+        _generateFutureDates(todo);
       });
+      _updateTodoList();
     }
   }
 
   // 반복 옵션 선택기
-  Future<void> _selectRepeatOption(BuildContext context, Todo todo) async {
+  void _selectRepeatOption(BuildContext context, Todo todo) async {
     String? selectedRepeat = await showDialog<String>(
       context: context,
-      builder: (context) {
+      builder: (BuildContext context) {
         return SimpleDialog(
           title: Text('반복 설정'),
-          children: [
-            SimpleDialogOption(child: Text('매일'), onPressed: () => Navigator.pop(context, '매일')),
-            SimpleDialogOption(child: Text('매주'), onPressed: () => Navigator.pop(context, '매주')),
-            SimpleDialogOption(child: Text('매월'), onPressed: () => Navigator.pop(context, '매월')),
-            SimpleDialogOption(child: Text('반복 안 함'), onPressed: () => Navigator.pop(context, null)),
+          children: <Widget>[
+            _buildRepeatOptionDialog('매일'),
+            _buildRepeatOptionDialog('매주'),
+            _buildRepeatOptionDialog('매월'),
+            _buildRepeatOptionDialog('반복 안함'),
           ],
         );
       },
     );
     if (selectedRepeat != null) {
       setState(() {
-        todo.repeat = selectedRepeat;
-        _updateTodoList();
+        todo.repeat = selectedRepeat == '반복 안함' ? null : selectedRepeat;
+        if (selectedRepeat == '매일') {
+          todo.repeatDays = ['월', '화', '수', '목', '금', '토', '일'];
+        } else if (selectedRepeat == '매주') {
+          todo.repeatDays = ['월', '화', '수', '목', '금', '토', '일'];
+          _selectWeekDays(context, todo);
+        } else if (selectedRepeat == '매월') {
+          todo.repeatDays = [];
+        } else if (selectedRepeat == '반복 안함') {
+          todo.repeatDays = [];
+        }
+        _generateFutureDates(todo);
       });
+      _updateTodoList();
     }
+  }
+
+  Future<void> _generateFutureDates(Todo todo) async {
+    if (todo.repeat == null || todo.alarmTime == null) {
+      todo.futureDates = null;
+      return;
+    }
+
+    List<Todo> futureTodos = [];
+    DateTime currentDate = todo.date;
+
+    for (int i = 0; i < 52; i++) { // 1년치 생성
+      switch (todo.repeat) {
+        case '매일':
+          currentDate = currentDate.add(Duration(days: 1));
+          break;
+        case '매주':
+          if (todo.repeatDays.isNotEmpty) {
+            do {
+              currentDate = currentDate.add(Duration(days: 1));
+            } while (!todo.repeatDays.contains(['월', '화', '수', '목', '금', '토', '일'][currentDate.weekday - 1]));
+          }
+          break;
+        case '매월':
+        // 다음 달의 같은 날짜로 설정
+          currentDate = DateTime(currentDate.year, currentDate.month + 1, currentDate.day);
+          // 만약 다음 달에 해당 날짜가 없다면 (예: 1월 31일 -> 2월 28일)
+          if (currentDate.day != todo.date.day) {
+            currentDate = DateTime(currentDate.year, currentDate.month, 0); // 해당 월의 마지막 날로 설정
+          }
+          break;
+        default:
+          return;
+      }
+
+      Todo newTodo = Todo(
+        title: todo.title,
+        date: currentDate,
+        memo: todo.memo,
+        alarmTime: todo.alarmTime,
+        repeat: todo.repeat,
+        repeatDays: todo.repeatDays,
+      );
+      futureTodos.add(newTodo);
+      await _addTodoToFirestore(newTodo); // 새로운 할 일을 Firestore에 추가
+    }
+
+    setState(() {
+      _todoList.addAll(futureTodos);
+    });
+    _updateTodoList();
   }
 
   // 옵션 박스 스타일링
   Widget _buildOptionBox({required IconData icon, required String text}) {
     return Container(
-      padding: EdgeInsets.all(10),
+      padding: EdgeInsets.symmetric(horizontal: 10, vertical: 8),
       decoration: BoxDecoration(
         color: Colors.white.withOpacity(0.1),
-        borderRadius: BorderRadius.circular(8),
+        borderRadius: BorderRadius.circular(5),
       ),
       child: Row(
+        mainAxisAlignment: MainAxisAlignment.center,
         children: [
-          Icon(icon, color: Colors.black),
-          SizedBox(width: 10),
+          Icon(icon, color: Colors.blue),
+          SizedBox(width: 8),
           Text(text),
         ],
       ),
@@ -388,12 +460,12 @@ class _TodoPageState extends State<TodoPage> {
   // 완료 버튼
   Widget _buildCompletionButton(Todo todo) {
     return ElevatedButton(
-      onPressed: () async {
+      onPressed: () async{
         setState(() {
           todo.isDone = !todo.isDone;
         });
         await _updateTodoInFirestore(todo);
-        _updateTodoList();
+        Navigator.pop(context);
       },
       child: Text(todo.isDone ? '완료 취소' : '완료하기'),
     );
@@ -401,26 +473,74 @@ class _TodoPageState extends State<TodoPage> {
 
   // 할 일 삭제 버튼
   Widget _buildDeleteButton(Todo todo) {
-    return ElevatedButton(
+    return ElevatedButton.icon(
       onPressed: () {
-        setState(() {
-          _todoList.remove(todo);
-        });
-        widget.onTodoRemoved(todo);  // 삭제된 할 일 콜백 호출
-        _deleteTodoFromFirestore(todo.id);
-        _updateTodoList();
-        Navigator.pop(context);
-      },
-      child: Text('삭제'),
-      style: ElevatedButton.styleFrom(
-        backgroundColor: Colors.red, ),
 
+        showDialog(
+          context: context,
+          builder: (BuildContext context) {
+            return AlertDialog(
+              title: Text('할 일 삭제'),
+              content: Text('삭제하시겠습니까?'),
+              actions: [
+                TextButton(
+                  child: Text('취소'),
+                  onPressed: () => Navigator.of(context).pop(),
+                ),
+                TextButton(
+                  child: Text('삭제'),
+                  onPressed: () async {
+                    await _deleteTodoFromFirestore(todo.id);
+                    setState(() {
+                      _todoList.remove(todo);
+                    });
+                    _updateTodoList();
+                    Navigator.of(context).pop();
+                    Navigator.of(context).pop();
+                  },
+                ),
+                if (todo.repeat != null)
+                  TextButton(
+                    child: Text('일정에서 전체 삭제'),
+                    onPressed: () async {
+                      // 데이터베이스에서 관련된 모든 일정 삭제
+                      List<Todo> relatedTodos = _todoList.where((t) =>
+                      t.title == todo.title &&
+                          t.repeat == todo.repeat &&
+                          t.alarmTime == todo.alarmTime
+                      ).toList();
+
+                      for (var relatedTodo in relatedTodos) {
+                        await _deleteTodoFromFirestore(relatedTodo.id);
+                      }
+
+                      setState(() {
+                        _todoList.removeWhere((t) =>
+                        t.title == todo.title &&
+                            t.repeat == todo.repeat &&
+                            t.alarmTime == todo.alarmTime
+                        );
+                      });
+                      _updateTodoList();
+                      Navigator.of(context).pop();
+                      Navigator.of(context).pop();
+                    },
+                  ),
+              ],
+            );
+          },
+        );
+      },
+      icon: Icon(Icons.delete, color: Colors.red),
+      label: Text('삭제'),
+      style: ElevatedButton.styleFrom(
+        backgroundColor: Colors.red.withOpacity(0.8),
+      ),
     );
   }
 
   // 새 할 일 추가 다이얼로그
-  // 새 할 일 추가 다이얼로그
-  void _addNewTodoDialog() {
+  void _addNewTodoDialog()  {
     TextEditingController _controller = TextEditingController();
 
     showDialog(
@@ -451,7 +571,7 @@ class _TodoPageState extends State<TodoPage> {
 
                   try {
                     // Firestore에 새로운 할 일 추가 후 setState 호출
-                    await _addTodoToFirestore(newTodo);  // Firestore에 추가
+                    await _addTodoToFirestore(newTodo); // Firestore에 추가
                     setState(() {
                       _todoList.add(newTodo); // 할 일 목록에 추가
                     });
@@ -476,6 +596,233 @@ class _TodoPageState extends State<TodoPage> {
   }
 
   bool isSameDay(DateTime date1, DateTime date2) {
-    return date1.year == date2.year && date1.month == date2.month && date1.day == date2.day;
+    return date1.year == date2.year && date1.month == date2.month &&
+        date1.day == date2.day;
+  }
+
+// 특정 날짜의 Todo 항목들을 반환하는 함수
+  List<Todo> _getEventsForDay(DateTime day) {
+    return _todoList.where((todo) =>
+    isSameDay(todo.date, day) ||
+        (todo.futureDates?.any((futureDate) => isSameDay(futureDate, day)) ??
+            false)
+    ).toList();
+  }
+
+// 미래 날짜의 Todo 항목들을 추가하는 함수
+  void _addFutureTodos(Todo originalTodo) {
+    if (originalTodo.futureDates == null) return;
+
+    for (DateTime futureDate in originalTodo.futureDates!) {
+      Todo newTodo = Todo(
+        title: originalTodo.title,
+        date: futureDate,
+        memo: originalTodo.memo,
+        alarmTime: originalTodo.alarmTime,
+        repeat: originalTodo.repeat,
+      );
+      setState(() {
+        _todoList.add(newTodo);
+      });
+    }
+    _updateTodoList();
+  }
+// Todo 페이지의 UI를 구성하는 build 메서드
+
+// Todo 항목의 반복 설정 텍스트를 반환하는 함수
+  String _getRepeatText(Todo todo) {
+    if (todo.repeat == null) return '반복 없음';
+    if (todo.repeat == '매일') return '매일';
+    if (todo.repeat == '매주') {
+      return '매주 ${todo.repeatDays.join(', ')}';
+    }
+    if (todo.repeat == '매월') return '매월';
+    return '반복 없음';
+  }
+
+// 반복 요일을 선택하는 옵션을 생성하는 위젯
+  Widget _buildRepeatDaysOption(Todo todo) {
+    return StatefulBuilder(
+        builder: (BuildContext context, StateSetter setState) {
+          return Wrap(
+            spacing: 8.0,
+            children: ['월', '화', '수', '목', '금', '토', '일'].map((day) {
+              return FilterChip(
+                label: Text(day),
+                selected: todo.repeatDays.contains(day),
+                onSelected: (bool selected) {
+                  setState(() {
+                    if (selected) {
+                      todo.repeatDays.add(day);
+                    } else {
+                      todo.repeatDays.remove(day);
+                    }
+                    if (todo.repeatDays.length == 7) {
+                      todo.repeat = '매일';
+                    } else if (todo.repeatDays.isNotEmpty) {
+                      todo.repeat = '매주';
+                    } else {
+                      todo.repeat = null;
+                    }
+                    _generateFutureDates(todo);
+                  });
+                  _updateTodoList();
+                },
+              );
+            }).toList(),
+          );
+        }
+    );
+  }
+  // 반복 설정을 위한 다이얼로그를 표시하는 함수
+  void _showRepeatOptionsDialog(BuildContext context, Todo todo) {
+    showDialog(
+      context: context,
+      builder: (BuildContext context) {
+        return StatefulBuilder(
+          builder: (BuildContext context, StateSetter setState) {
+            return AlertDialog(
+              title: Text('반복 설정'),
+              content: _buildRepeatCheckboxes(todo, setState), // 반복 체크박스 UI 생성
+              actions: <Widget>[
+                TextButton(
+                  child: Text('취소'),
+                  onPressed: () => Navigator.of(context).pop(), // 다이얼로그 닫기
+                ),
+                TextButton(
+                  child: Text('확인'),
+                  onPressed: () {
+                    _updateTodoList(); // Todo 리스트 업데이트
+                    Navigator.of(context).pop(); // 다이얼로그 닫기
+                  },
+                ),
+              ],
+            );
+          },
+        );
+      },
+    );
+  }
+
+// 반복 설정을 위한 체크박스 UI를 생성하는 함수
+  Widget _buildRepeatCheckboxes(Todo todo, StateSetter setState) {
+    return Column(
+      mainAxisSize: MainAxisSize.min,
+      children: [
+        CheckboxListTile(
+          title: Text('반복 안함'),
+          value: todo.repeat == null,
+          onChanged: (bool? value) {
+            setState(() {
+              todo.repeat = value! ? null : '매주';
+              todo.repeatDays = [];
+            });
+          },
+        ),
+        CheckboxListTile(
+          title: Text('매일'),
+          value: todo.repeat == '매일',
+          onChanged: (bool? value) {
+            setState(() {
+              todo.repeat = value! ? '매일' : null;
+              todo.repeatDays =
+              value ? ['월', '화', '수', '목', '금', '토', '일'] : [];
+            });
+          },
+        ),
+        CheckboxListTile(
+          title: Text('매주'),
+          value: todo.repeat == '매주',
+          onChanged: (bool? value) {
+            setState(() {
+              todo.repeat = value! ? '매주' : null;
+              if (!value) todo.repeatDays = [];
+            });
+          },
+        ),
+        if (todo.repeat == '매주')
+          Wrap(
+            spacing: 8.0,
+            children: ['월', '화', '수', '목', '금', '토', '일'].map((day) {
+              return FilterChip(
+                label: Text(day),
+                selected: todo.repeatDays.contains(day),
+                onSelected: (bool selected) {
+                  setState(() {
+                    if (selected) {
+                      todo.repeatDays.add(day);
+                    } else {
+                      todo.repeatDays.remove(day);
+                    }
+                    if (todo.repeatDays.isEmpty) {
+                      todo.repeat = null;
+                    }
+                  });
+                },
+              );
+            }).toList(),
+          ),
+        CheckboxListTile(
+          title: Text('매월'),
+          value: todo.repeat == '매월',
+          onChanged: (bool? value) {
+            setState(() {
+              todo.repeat = value! ? '매월' : null;
+              todo.repeatDays = [];
+            });
+          },
+        ),
+      ],
+    );
+  }
+  void _selectWeekDays(BuildContext context, Todo todo) {
+    showDialog(
+      context: context,
+      builder: (BuildContext context) {
+        return StatefulBuilder(
+          builder: (BuildContext context, StateSetter setState) {
+            return AlertDialog(
+              title: Text('요일 선택'),
+              content: Wrap(
+                spacing: 5,
+                children: ['월', '화', '수', '목', '금', '토', '일'].map((day) {
+                  return FilterChip(
+                    label: Text(day),
+                    selected: todo.repeatDays.contains(day),
+                    onSelected: (bool selected) {
+                      setState(() {
+                        if (selected) {
+                          todo.repeatDays.add(day);
+                        } else {
+                          todo.repeatDays.remove(day);
+                        }
+                      });
+                    },
+                  );
+                }).toList(),
+              ),
+              actions: <Widget>[
+                TextButton(
+                  child: Text('확인'),
+                  onPressed: () {
+                    Navigator.of(context).pop();
+                    _generateFutureDates(todo);
+                  },
+                ),
+              ],
+            );
+          },
+        );
+      },
+    );
+  }
+
+  SimpleDialogOption _buildRepeatOptionDialog(String option) {
+    return SimpleDialogOption(
+      onPressed: () {
+        Navigator.pop(context, option);
+      },
+      child: Text(option),
+    );
   }
 }
